Code/requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-appconf==1.0.4
django-imagekit==4.0.2
django-mdeditor==0.1.18
django-redis==4.12.1
djangorestframework==3.12.4
idna==2.10
importlib-metadata==4.0.1
itypes==1.2.0
Jinja2==2.11.3
Markdown==3.3.4
MarkupSafe==1.1.1
mysqlclient==2.0.3
pilkit==2.0
Pillow==8.2.0
Pygments==2.8.1
pymdown-extensions==8.1.1
python-markdown-math==0.8
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
uwsgi
zipp==3.4.1
